[mobile][locker] Tighten the scanner's Dart side
Three unrelated bits of slack:

ScanQuad carried its corners in the segmentation mask's 256x256 space, so
fourteen sites multiplied or divided by ScanQuad.maskSide to get anywhere
useful. Corners are fractions of the frame now, which is what every consumer
actually wanted: the overlay scales straight to its canvas, the crop page
keeps its normalised handles as they are, and the two area checks compare
against a fraction without rescaling. The mask side survives as one private
constant in the service, where the live-detection quad crosses the FFI
boundary in mask space.

DocumentScannerService was an abstract class with one implementation reached
through a one-line service locator, and nothing consumed the injection seam.
It is now a single class with the usual .instance singleton, in the file that
already had its name.

buildPdf parsed every page's JPEG twice: once for the page size, once inside
PdfWriter. ScannedPage already carries the processed dimensions.

flutter analyze clean over the whole app; scratch tests confirmed the
normalised area and drift thresholds behave as before. -91 lines, 10 files
instead of 12.
